idk what changed lol

+++ src/map/block.cs
@@ -1,3 +1,3 @@
 public class block {
-    public byte index;
+    public ushort index;
 }
